Updated testing for pfp and user profile
User profile is now able to be pulled from data base using "/user/"UserId:"

Replace :UserId with any of the already made users, tested it with my current code and works 100% to pull the user's info, such as pfp, score, and name and email. Currently integrating the listing dummy data and encourporating a way for users to change their pfp at any time, and have is saved on the database.