Updating User Profile and Msg option
Currently added the ability for user to send msg and store within firebase in order to pull and return written information. This shall be store within the db and can be retrieved when in the msg board of certain user. Will have to add on to make it work in real time and work without redirecting the user back to their own profiile. Couple bug fixes which shouldnt take too long.