Added Gemini AI API Feature
Currently working on improving the site with a AI Assistant using the Gemini API, will create short and consise replies when signed in, will only open when signed in ( Could make it so that a pop up comes up telling users to log in to use). Currently will save conversations to the database but under "Users" will have to tweak to make it save under seperate collection, kinda still working with making everything work in cinque. Will add photo feature next pr.